+++ Pack/Release-Notes.md
@@ -2,6 +2,11 @@
 PowerShelf Release Notes
 ========================
 
+## v1.1.0
+
+New script *Assert-SameFile.ps1* automates one typical test scenario, it
+compares the sample and result files and performs copy and view operations.
+
 ## v1.0.0
 
 NuGetDebugTools [v1.3.0](https://github.com/nightroman/PowerShelf/blob/master/Pack/NuGetDebugTools/Release-Notes.md#v130)